requirements.txt
Flask>=3.0.3
Werkzeug>=3.0.3
Jinja2>=3.1.4
itsdangerous>=2.2.0
click>=8.1.7

Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.37

cryptography>=42.0.8
python-dotenv>=1.0.1
groq>=0.9.0

gunicorn>=22.0.0

# Áudio / Vídeo
faster-whisper==1.0.3
ctranslate2==4.4.0
av==12.3.0
numpy>=1.26.0
ffmpeg-python==0.2.0

# OCR (IMAGEM)
pytesseract>=0.3.10
Pillow>=10.4.0
opencv-python>=4.10.0.84
pip install pytesseract pillow opencv-python
